# migrations/openapi.yaml
openapi: 3.0.0
info:
  title: Learning Backend API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/courses:
    get:
      summary: 课程列表（课程入口）
      description: 用于多课程导览入口，支持按关键词搜索课程。
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: OK }
...
  /api/tasks:
    get:
      summary: 任务列表（课程内）
      description: 按课程列出任务清单，并附带当前学生最近一次提交情况，用于课程任务页。
      parameters:
        - in: query
          name: course_id
          required: true
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: OK }
...